pyproject.toml
[project]
name = "telegram-bot-gastos-llm"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
    "python-telegram-bot==21.0.1",
    "gspread>=6.1.2",
    "google-auth>=2.29.0",
    "ollama",
    "python-dotenv>=1.0.1",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "black>=24.3.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
